requirements/common.txt
Django==1.4.3
django-autocomplete_light==1.1.30
fdb==1.0